Movi la funcion de formatear el la fecha del curso a utils
